refactor: 출석기한 별도 클래스로 분리
