+++ src/main/java/org/cotato/csquiz/api/attendance/dto/UpdateAttendanceRequest.java
@@ -1,7 +1,7 @@
 package org.cotato.csquiz.api.attendance.dto;
 
-import static org.cotato.csquiz.domain.attendance.constant.DeadLineUtil.DEFAULT_END_TIME;
-import static org.cotato.csquiz.domain.attendance.constant.DeadLineUtil.DEFAULT_START_TIME;
+import static org.cotato.csquiz.domain.attendance.constant.DeadLineConstants.DEFAULT_END_TIME;
+import static org.cotato.csquiz.domain.attendance.constant.DeadLineConstants.DEFAULT_START_TIME;
 
 import jakarta.validation.constraints.NotNull;
 import java.util.Objects;
