+++ src/main/java/org/cotato/csquiz/domain/attendance/constant/DeadLineConstants.java
@@ -2,7 +2,7 @@ package org.cotato.csquiz.domain.attendance.constant;
 
 import java.time.LocalTime;
 
-public class DeadLineUtil {
+public class DeadLineConstants {
     public static final LocalTime DEFAULT_START_TIME = LocalTime.of(19, 0);
     public static final LocalTime DEFAULT_END_TIME = LocalTime.of(19, 20);
 }
